+++ CHANGELOG.md
@@ -8,6 +8,11 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 #### Major Changes
 1. Extended support to Server Profile resource on DL Server.
+2. Extended support to Server Profile Template resource on DL Server.
+3. Extended support to Server Hardware resource for DL Server.
+   - POST   /rest/server-hardware
+   - PATCH  /rest/server-hardware/{id}
+4. logical_interconnect_group: Changed the way we provide port_num, instead of list we have to provide integer value for port_num in each logical_port_config in uplink_set.
 
 # [v6.2.0-13]